ep1: Accept a customer's order data - Updated purchaseTotal to Float for rewardPoints Float in orderRequest

+++ src/main/kotlin/com/urbn/rewards/models/Customer.kt
@@ -2,7 +2,7 @@ package com.urbn.rewards.models
 
 data class Customer(
     val email: String,
-    val rewardPoints: Int,
+    val rewardPoints: Float,
     val rewardsTier: String,
     val nextRewardsTier: String,
     val nextRewardsTierName: String,
